+++ core/src/commonMain/kotlin/NavigationChain.kt
@@ -1,16 +1,16 @@
 package dev.inmo.navigation.core
 
 import dev.inmo.kslog.common.*
+import dev.inmo.micro_utils.common.diff
 import dev.inmo.micro_utils.coroutines.*
 import dev.inmo.navigation.core.extensions.*
-import dev.inmo.navigation.core.visiter.walk
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 
 class NavigationChain<Base>(
-    internal val parentNode: NavigationNode<out Base, Base>?,
+    val parentNode: NavigationNode<out Base, Base>?,
     internal val nodeFactory: NavigationNodeFactory<Base>,
     val id: NavigationChainId? = null
 ) {
@@ -73,38 +73,56 @@ class NavigationChain<Base>(
         }
     }
 
-    fun drop(id: NavigationNodeId): NavigationNode<out Base, Base>? {
-        val node = nodesIds[id] ?: return null
-        node.state = NavigationNodeState.NEW
-        log.d { "Stack before removing of $node: ${stackFlow.value.joinToString("; ") { it.toString() }}" }
-        _stackFlow.value = _stackFlow.value.filterNot { currentNode ->
-            (currentNode.id == id).also {
-                log.d { "Id of $currentNode (${currentNode.id}) is equal to $id: $it" }
+    fun drop(node: NavigationNode<*, Base>): NavigationNode<out Base, Base>? {
+        var dropped = false
+
+        val newStack = stack.filterNot { currentNode ->
+            (currentNode === node).also {
+                dropped = true
+                log.d { "$currentNode (${currentNode.id}) is equal to ${node}: $it" }
             }
         }
+        if (!dropped) {
+            return null
+        }
+
+        node.state = NavigationNodeState.NEW
+
+        log.d { "Stack before removing of $node: ${stackFlow.value.joinToString("; ") { it.toString() }}" }
+        _stackFlow.value = newStack
         log.d { "Stack after removing of $node: ${stackFlow.value.joinToString("; ") { it.toString() }}" }
-        nodesIds.remove(id)
+
+        nodesIds.remove(node.id)
         return node
     }
+
+    fun drop(id: NavigationNodeId): NavigationNode<out Base, Base>? {
+        return drop(nodesIds[id] ?: return null)
+    }
     fun drop(id: String) = drop(NavigationNodeId(id))
+
+    fun replace(
+        node: NavigationNode<*, Base>,
+        config: Base
+    ): Pair<NavigationNode<out Base, Base>, NavigationNode<out Base, Base>>? {
+        val i = stack.indexOfFirst { it === node }.takeIf { it > -1 } ?: return null
+
+        nodesIds.remove(node.id)
+        node.state = NavigationNodeState.NEW
+
+        val newNode = nodeFactory.createNode(this, config) ?: return null
+
+        _stackFlow.value = (stack.take(i) + newNode) + stack.drop(i + 1)
+        nodesIds[newNode.id] = newNode
+
+        return node to newNode
+    }
 
     fun replace(
         id: NavigationNodeId,
         config: Base
     ): Pair<NavigationNode<out Base, Base>, NavigationNode<out Base, Base>>? {
-        val i = stack.indexOfFirst { it.id == id }.takeIf { it > -1 } ?: return null
-
-        val newNode = nodeFactory.createNode(this, config) ?: return null
-        val oldNode = stack[i]
-        val currentStack = _stackFlow.value
-        _stackFlow.value = currentStack.take(i) + newNode + currentStack.drop(i + 1)
-
-        nodesIds.remove(id)
-        nodesIds[newNode.id] = newNode
-
-        oldNode.state = NavigationNodeState.NEW
-
-        return oldNode to newNode
+        return replace(stack.firstOrNull { it.id == id } ?: return null, config)
     }
 
     fun replace(
@@ -116,6 +134,9 @@ class NavigationChain<Base>(
         while (stack.isNotEmpty()) {
             pop()
         }
+    }
+    fun dropItself(): Boolean {
+        return parentNode ?.removeSubChain(this) == true
     }
 
     @Deprecated("Extracted to the walking API", ReplaceWith("this.dropInSubTree(id)", "dev.inmo.navigation.core.extensions.dropInSubTree"))
@@ -129,11 +150,11 @@ class NavigationChain<Base>(
         config: Base
     ) = replaceInSubTree(id, config)
 
-    @Deprecated("Extracted to the walking API", ReplaceWith("this.replaceInSubTree(id, config)", "dev.inmo.navigation.core.extensions.replaceInSubTree"))
+    @Deprecated("Extracted to the walking API", ReplaceWith("this.replaceNodesInSubTree(id, config)", "dev.inmo.navigation.core.extensions.replaceNodesInSubTree"))
     fun replaceInTree(
         id: String,
         config: Base
-    ) = replaceInSubTree(id, config)
+    ) = replaceNodesInSubTree(id, config)
 
     @Deprecated("Extracted to the walking API", ReplaceWith("this.pushInSubTree(id, config)", "dev.inmo.navigation.core.extensions.pushInSubTree"))
     fun pushInTree(
@@ -141,11 +162,11 @@ class NavigationChain<Base>(
         config: Base
     ) = pushInSubTree(id, config)
 
-    @Deprecated("Extracted to the walking API", ReplaceWith("this.pushInSubTreeByNodeId(id, config)", "dev.inmo.navigation.core.extensions.pushInSubTreeByNodeId"))
+    @Deprecated("Extracted to the walking API", ReplaceWith("this.pushInNodesInSubTree(id, config)", "dev.inmo.navigation.core.extensions.pushInNodesInSubTree"))
     fun pushInTree(
         id: String,
         config: Base
-    ) = pushInSubTreeByNodeId(id, config)
+    ) = pushInNodesInSubTree(id, config)
 
     fun start(scope: CoroutineScope): Job {
         val subscope = scope.LinkedSupervisorScope()
@@ -165,38 +186,30 @@ class NavigationChain<Base>(
             log.d { "Cancelled subscope" }
         }
 
-        stackFlow.subscribeSafelyWithoutExceptions(subscope) {
-            actualizeStackStates()
-        }
-
         val nodeToJob = mutableMapOf<NavigationNodeId, Job>()
-        val nodeToJobMutex = Mutex()
 
-        (onNodeAddedFlow + onNodeReplacedFlow.map { it.map { it.second } }).flatten().subscribeSafelyWithoutExceptions(subscope) {
-            nodeToJobMutex.withLock {
-                nodeToJob[it.value.id] = it.value.start(subscope)
+        merge(
+            flow { emit(emptyList<NavigationNode<*, Base>>().diff(stackFlow.value)) },
+            onNodesStackDiffFlow
+        ).subscribeSafelyWithoutExceptions(subscope) {
+            it.removed.forEach { (i, it) ->
+                nodeToJob.remove(it.id) ?.cancel()
             }
-        }
-        (onNodeRemovedFlow + onNodeReplacedFlow.map { it.map { it.first } }).flatten().subscribeSafelyWithoutExceptions(subscope) {
-            nodeToJobMutex.withLock {
-                nodeToJob.remove(it.value.id) ?.cancel()
+            it.replaced.forEach { (old, new) ->
+                nodeToJob[new.value.id] = new.value.start(subscope)
+                nodeToJob.remove(old.value.id) ?.cancel()
             }
-        }
-        onNodeRemovedFlow.dropWhile { stack.isNotEmpty() }.subscribeSafelyWithoutExceptions(subscope) {
-            log.d { "Dropping myself from parent node $parentNode" }
-            parentNode ?.removeChain(this)
-            log.d { "Dropped myself from parent node $parentNode" }
-        }
+            it.added.forEach { (i, it) ->
+                nodeToJob[it.id] = it.start(subscope)
+            }
 
-        subscope.launch {
-            stackFlow.value.forEach {
-                nodeToJobMutex.withLock {
-                    nodeToJob[it.id] = it.start(subscope)
-                }
-            }
             actualizeStackStates()
-        }
 
+            if (stack.isEmpty()) {
+                dropItself()
+            }
+        }
+        
         return subscope.coroutineContext.job
     }
 }